+++ Packs/ServiceNow/Integrations/ServiceNowv2/test_data/result_constants.py
@@ -286,3 +286,8 @@ EXPECTED_GET_TABLE_NAME = {
         }
     ]
 }
+EXPECTED_ADD_TAG = {
+    'ServiceNow.Ticket(val.ID===obj.ID)': {
+        'ID': '123', 'TagTitle': 'title', 'TagID': '1234'
+    }
+}